requirements.txt
pyserial
numpy
pandas
matplotlib